Use M_PI instead of M_PIl
M_PIl is the double precision constant for pi (3.14...).
This is not defined in all toolchains, resulting in the
following build error:

```
test/testcpp.cc:46:35: error: use of undeclared identifier 'M_PIl'; did you mean 'P_PID'?
     46 |         long double phinc = 2*k0* M_PIl / nfft;
        |                                   ^~~~~
        |                                   P_PID
```

The double precision constants may be a GNU extension. See:
https://www.gnu.org/software/libc/manual/html_node/Mathematical-Constants.html_node

Switching the test to use the single precision M_PI avoids this
error. The test still passes as the added precision of the
double is not required.

This fixes #87.